requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
celery[redis]
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
click8==8.0.1
dnspython==1.16.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.0
Flask-RESTful==0.3.7
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==1.5.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
mysqlclient==1.4.6
pip==20.0.2
protobuf==3.6.1
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sentry-sdk[flask]==0.14.1
setuptools==45.1.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==0.16.0
wheel==0.33.6
zipp==2.2.0
